API Structure re-write
After a year or so with actual industry experience and a lot of consideration, I've re-designed the architecture from the ground-up, utilizing databases correctly and structuring the API a little more logically. I've also opted to use swagger instead of trying to brute-force my way through the process with flask alone. Much of the functionality is currently non-existent, but is in the process of being re-implemented.