resolved circular dependencies
__init__.py:
unneeded as of py 3.3+; removed

config.py, Dock.py:
moved adding the API away from the config file and into the main file, as it was causing a circular dependency (and is not the job of the config file anyway).

api-config.yaml, files.py:
added in a test API query, that gets the top x files, given a specific offset. This will likely change in the future as functionality is structured, but I needed to get a proof-of-concept working to test things out.